Fix static file serving path when using copyPath
The static files were served from the non-public path of project root prior to this change. Adding `contentBase` with `PUBLIC_DIRECTORY` path as the value to the webpack-dev-server configuration makes the static files be served relative to the public path, not the project root. The static file path prior to change, with `public/` as the public directory: `/public/copiedFolder/file.jpg`. After the change: `/copiedFolder/file.jpg`.