[BREAKING] Move universal-react server entry point to create-frontend
Before, the entry point lived in the user's app directory. It was an ugly file, containing hot reload logic and error handling. Ideally, the user shouldn't have to see that, or deal with that file. It also limited our ability to add in custom stuff like auto reloading .env variables.

The new solution creates a virtual module that acts as the server entry point, which then imports the actual user entry. The vmodule is necessary because we need to import the user entry, but that path comes from a variable that is not known before building.